Move JSONRequest and JSONResponse to new files
They're separate concepts and should be sent elsewhere.
We're going to be adding additional subclasses to that as well.
