+++ src/BoardClasses/ActiveGame.ts
@@ -1,174 +1,10 @@
-import { Board, Color } from "./Board"
-
-abstract class JSONRequest {
-    constructor() {}
-
-    isStateRequest(): boolean {
-        return false;
-    }
-    isMoveRequest(): boolean {
-        return false;
-    }
-    getMove(): [number, number] | null {
-        return null;
-    }
-}
-
-class JSONStateRequest extends JSONRequest {
-    constructor() { super(); }
-    isStateRequest(): boolean {
-        return true;
-    }
-}
-
-class JSONMoveRequest extends JSONRequest {
-    source: number;
-    target: number;
-
-    constructor(source: number, target: number) { 
-        super();
-        this.source = source;
-        this.target = target;
-    }
-
-    isMoveRequest(): boolean {
-        return true;
-    }
-
-    getMove(): [number, number] | null {
-        return [this.source, this.target];
-    }
-}
-
-abstract class JSONResponse {
-    constructor() {}
-
-    isInvalidMessage(): boolean {
-        return false;
-    }
-
-    isInvalidMove(): boolean {
-        return false;
-    }
-
-    isBoardState(): boolean {
-        return false;
-    }
-
-    isValidMove(): boolean {
-        return false;
-    }
-
-    isNotYourTurn(): boolean {
-        return false;
-    }
-
-    abstract toObject(): object;
-}
-
-class JSONInvalidMessageResponse extends JSONResponse {
-    message: string;
-
-    constructor(message: string) {
-        super();
-        this.message = message;
-    }
-
-    isInvalidMessage(): boolean {
-        return true;
-    }
-
-    toObject(): object {
-        return {
-            response_type: "invalid_message",
-            message: this.message
-        };
-    }
-}
-
-class JSONInvalidMoveResponse extends JSONResponse {
-    constructor() {
-        super();
-    }
-
-    isInvalidMove(): boolean {
-        return true;
-    }
-
-    toObject(): object {
-        return {
-            response_type: "invalid_move"
-        }
-    }
-}
-
-class JSONBoardStateResponse extends JSONResponse {
-    board: Board;
-
-    constructor(board: Board) {
-        super();
-        this.board = board.copy();
-    }
-
-    isBoardState(): boolean {
-        return true;
-    }
-
-    toObject(): object {
-        return {
-            response_type: "board_state",
-            board_state: this.board.toObject()
-        };
-    }
-}
-
-class JSONValidMoveResponse extends JSONResponse {
-    board: Board;
-
-    constructor(board: Board) {
-        super();
-        this.board = board.copy();
-    }
-
-    isValidMove(): boolean {
-        return true;
-    }
-
-    toObject(): object {
-        return {
-            response_type: "valid_move",
-            board_state: this.board.toObject()
-        };
-    }
-}
-
-class JSONIsNotYourTurnResponse extends JSONResponse {
-    constructor() { super(); }
-
-    isNotYourTurn(): boolean {
-        return true;
-    }
-
-    toObject(): object {
-        return {
-            response_type: "not_your_turn"
-        };
-    }
-}
-
-function parseMessageJSON(json: string): JSONRequest | null {
-    let parsedObj: any = JSON.parse(json);
-    if(parsedObj?.request_type && parsedObj.request_type === "get_state") {
-        return new JSONStateRequest();
-    }
-    if(parsedObj?.request_type && parsedObj.request_type === "move") {
-        if(Number.isInteger(parsedObj?.move?.source) &&
-           Number.isInteger(parsedObj?.move?.target)) {
-            return new JSONMoveRequest(parsedObj.move.source, parsedObj.move.target);
-        }
-    }
-    return null;
-}
+import { Board, Color } from "./Board";
+import { JSONRequest, parseMessageJSON } from '../JSONClasses/JSONRequest';
+import { JSONInvalidMessageResponse, 
+         JSONBoardStateResponse,
+         JSONIsNotYourTurnResponse,
+         JSONInvalidMoveResponse,
+         JSONValidMoveResponse } from '../JSONClasses/JSONResponse';
 
 export class ActiveGame {
     blackID: string;
